refactor(server): split http.go into per-responsibility files
What:
- split internal/server/http.go (2469 lines) into 12 files within
  the same `server` package:
  - http.go            core: struct, NewHTTPServer(routing), setters, lifecycle
  - http_interfaces.go querier/manager/writer interfaces (23)
  - http_ingest.go     OTLP receive: traces/metrics/logs
  - http_query.go      REST query/dashboard/SSE/raw + topology cache
  - http_logs.go       Log Analytics handlers
  - http_alerts.go     alert routing/history/ack
  - http_rca.go        deploy event + RCA reports
  - http_profiling.go  profiling + k8s metrics + infra correlation
  - http_slo.go        SLO definitions/burn-rate
  - http_catalog.go    service catalog
  - http_ops.go        stats/healthz/readyz
  - http_util.go       shared helpers
- pure move: no logic, signature, or API change
- relocated 4 orphaned section-divider comments to their group files

Why:
- 2469-line file hurt navigation/review; per-responsibility files
  make the ingest/query/log/alert/rca handlers easier to work on
- same recipe as clickhouse.go split (PR #21)

Verify:
- go build ./... / go vet ./... / gofmt -l . (clean)
- go test -race -count=1 ./... (all pass)
- 109 top-level declarations identical to original (diff)
- 2178 non-blank body lines byte-identical to original (diff)
